% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesdm_data.R
\docType{data}
\name{climate_data}
\alias{climate_data}
\title{Climate Data}
\format{
\subsection{\code{climate_data} A tibble 11 columns:}{
\itemize{
\item \code{climate_scenario}: Character. The climate scenario. Valid values:
\code{current}, \code{ssp126}, \code{ssp370}, \code{ssp585}.
\item \code{climate_model}: Character. The climate model. Valid values: \code{current},
\code{GFDL-ESM4}, \code{IPSL-CM6A-LR}, \code{MPI-ESM1-2-HR}, \code{MRI-ESM2-0}, and
\code{UKESM1-0-LL}.
\item \code{climate_model_abb}: Character. The abbreviated climate model name. Valid
values are: \code{current}, \code{gfdl}, \code{ipsl}, \code{mpi}, \code{mri}, and \code{ukesm1}.
\item \code{year}: Character. The year ranges. Valid values are: \verb{1981_2010},
\verb{2011_2040}, \verb{2041_2070}, and \verb{2071_2100}.
\item \code{var_name}: Character. The variable name. Valid values: \code{bio1}, \code{bio2},
\code{bio3}, \code{bio4}, \code{bio5}, \code{bio6}, \code{bio7}, \code{bio8}, \code{bio9}, \code{bio10}, \code{bio11},
\code{bio12}, \code{bio13}, \code{bio14}, \code{bio15}, \code{bio16}, \code{bio17}, \code{bio18}, \code{bio19},
\code{fcf}, \code{fgd}, \code{gdd0}, \code{gdd5}, \code{gdd10}, \code{gddlgd0}, \code{gddlgd5}, \code{gddlgd10},
\code{gdgfgd0}, \code{gdgfgd5}, \code{gdgfgd10}, \code{gsl}, \code{gsp}, \code{gst}, \code{lgd}, \code{ngd0},
\code{ngd5}, \code{ngd10}, \code{npp}, \code{scd}, and \code{swe}.
\item \code{climate_name}: Character. The climate name (e.g., \code{current} and
<\code{year}>\emph{<\code{climate_scenario}>}<\code{climate_model_abb}> for future options (all
small letters)).
\item \code{osf_path}: Character. The OSF path to the climate data file (e.g.,
"1981_2010_res_10/bio1.tif").
\item \code{file_size_mb}: Double. The size of the climate data file in megabytes.
\item \code{resolution}: Integer. The resolution of the climate data file. Valid
values are \code{5}, \code{10}, and \code{20}.
\item \code{out_dir}: Character. The output directory for the climate data file.
e.g. ("\verb{res_5/1981_2010}"). This would be relative to the \code{climate_dir}
argument.
\item \code{out_file}: Character. The output file path for the climate data file.
e.g. ("\verb{res_5/1981_2010/bio1.tif}"). This would be relative to the
\code{climate_dir} argument.
}
}
}
\usage{
climate_data
}
\description{
A dataset containing information about climate data files available for
download from OSF.
}
\examples{
#' # View the climate data
data("climate_data")
head(climate_data)
}
\keyword{datasets}
