% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_osf.R
\name{get_landuse_data}
\alias{get_landuse_data}
\title{Download Land Use Data from OneSDM OSF Project}
\usage{
get_landuse_data(
  climate_dir = NULL,
  resolution = 10L,
  climate_scenario = "current",
  year = "1981_2010",
  pft_type = "cross-walk",
  pft_id = NULL,
  verbose = TRUE,
  sleep_time = 1L
)
}
\arguments{
\item{climate_dir}{Character. Destination directory where climate and land
use data files will be saved. Can be set via the "\code{onesdm_climate_dir}"
option. The same directory should be used in case of modelling multiple
species to ensure consistency. Default is \code{NULL}.}

\item{resolution}{Numeric. Spatial resolution. Valid values are 5, 10, or 20
for resolutions of approximately 5, 10, and 20 km (2.5, 5, and 10
arc-minutes) respectively. Can be set via the "\code{onesdm_resolution}" option.
Default is \code{10L}.}

\item{climate_scenario}{Character scalar. Climate scenario; one of \code{current},
\code{ssp126}, \code{ssp370}, \code{ssp585}. Default is \code{"current"}.}

\item{year}{Character scalar. Time period; one of \verb{1981_2010}, \verb{2011_2040},
\verb{2041_2070}, \verb{2071_2100}. Default is \code{"1981_2010"}.}

\item{pft_type}{Character string. Plant functional type category. Has to be
one of "\code{cross-walk}" (default) "\code{original}"; see \link{landuse_data}
for details. This can be set via the "\code{onesdm_pft_type}" option.}

\item{pft_id}{Numeric vector. One or more plant functional type identifiers
to download. Must be valid for the specified \code{pft_type}: 1-20 for \code{pft_type == "original"}, and 1-12 for \code{pft_type == "cross-walk"}; See
\link{landuse_data} for details. Can be set via the "\code{onesdm_pft_id}"
option. This parameter has no default and must be provided.}

\item{verbose}{Logical scalar. If \code{TRUE} (default), prints progress and
informative messages.}

\item{sleep_time}{Numeric scalar. Number of seconds to pause after
downloading files to avoid overwhelming the server. Default is \code{1L}.}
}
\value{
Invisibly returns a tibble containing metadata about the downloaded
files, including local file paths, OSF paths, download links, and
validation status.
}
\description{
Downloads selected land use data files from the OneSDM Open
Science Framework (OSF) storage based on spatial resolution, climate
scenario, climate model, time period, and Plant Functional Type (PFT)
names. The function validates input parameters, checks for existing files,
and downloads only missing or corrupted files.
}
\details{
The function performs the following steps:
\itemize{
\item Validates all input parameters against available options
\item Checks if requested files already exist locally and are valid
\item Creates necessary output directories
\item Connects to the OneSDM OSF project and downloads only missing or corrupted
files
\item Validates downloaded files using TIFF format checks
}
}
\examples{
\dontrun{
  ecokit::load_packages(fs, terra)

  # Create a temporary directory for land use data
  tmp_dir <- fs::path_temp("onesdm_landuse")
  fs::dir_create(tmp_dir)

  # |||||||||||||||||||||||||

  # Example usage of get_landuse_data
  get_landuse_data(
    climate_dir = tmp_dir,
    resolution = 20L,
    climate_scenario = "current",
    year = "1981_2010",
    pft_type = "cross-walk",
    pft_id = c(2, 4, 12),
    verbose = TRUE)

  # List downloaded files in the subdirectory of combination of parameters
  print(list.dirs(tmp_dir))

  list_files <- list.files(
    fs::path(tmp_dir, "res_20", "1981_2010"), full.names = TRUE)
  print(list_files)

  # Load the downloaded files as a SpatRaster stack
  terra::rast(list_files)

  # |||||||||||||||||||||||||

  # Example using future land use data set using options to set defaults

  options(
    onesdm_climate_dir = tmp_dir, onesdm_resolution = 20L,
    onesdm_verbose = TRUE,
    onesdm_pft_type = "original",
    onesdm_pft_id = c(1, 14, 18))

  get_landuse_data(climate_scenario = "ssp585", year = "2071_2100")

  # List downloaded files in the subdirectory of combination of parameters
  print(list.dirs(tmp_dir))

  # Load the downloaded files as a SpatRaster stack
  list_files_future <- list.files(
    fs::path(tmp_dir, "res_20", "2071_2100_ssp585"),
    recursive = TRUE, pattern = "\\\\.tif$", full.names = TRUE)

  print(list_files_future)

  # Load the downloaded files as a SpatRaster stack
  terra::rast(list_files_future)

  # |||||||||||||||||||||||||

  # Clean up temporary directory after use
  fs::dir_delete(tmp_dir)
}

}
\author{
Ahmed El-Gabbas
}
