% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_osf.R
\name{load_mask_layer}
\alias{load_mask_layer}
\title{Load or Create a Mask Layer at Specified Resolution}
\usage{
load_mask_layer(
  resolution = NULL,
  climate_dir = NULL,
  verbose = FALSE,
  overwrite = FALSE,
  return_spatraster = TRUE,
  wrap = FALSE
)
}
\arguments{
\item{resolution}{Integer. The aggregation factor or resolution for the mask
layer. Valid options are: \code{5L}, \code{10L}, or \code{20L} for resolutions of ~ 5km, ~
10km, or ~ 20km respectively. Default is \code{10L}.}

\item{climate_dir}{Character. Directory where climate data and mask layers
are stored. If \code{NULL} (default), a temporary file is created.}

\item{verbose}{Logical. Should a download progress bar be shown? Default is
\code{FALSE}.}

\item{overwrite}{Logical. Should existing files be overwritten? Default is
\code{FALSE}.}

\item{return_spatraster}{Logical. If \code{TRUE} (default), returns a \code{SpatRaster}
object. If \code{FALSE}, saves the raster to a file and returns the file path
(invisibly).}

\item{wrap}{Logical. Should the resulting \code{SpatRaster} be wrapped using
\link[terra:wrap]{terra::wrap}? Default is \code{FALSE}.}
}
\value{
A \code{SpatRaster} object representing the mask layer at the specified
resolution, or the file path (invisibly) if \code{return_spatraster} is \code{FALSE}.
}
\description{
Loads a mask layer (raster) at a specified model resolution from the OneSDM
Open Science Framework (OSF) storage. It can either return a \code{SpatRaster}
object directly or saved to a specified path.
}
\examples{
\dontrun{
# Load mask layer as `SpatRaster` object
mask <- load_mask_layer(resolution = 10L)

# Save mask layer to file
load_mask_layer(resolution = 20L, climate_dir = NULL)
}

}
\author{
Ahmed El-Gabbas
}
