% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_climate.R
\name{prepare_climate}
\alias{prepare_climate}
\title{Prepare CHELSA Climate Data}
\usage{
prepare_climate(
  agg_factors = c(5L, 10L, 20L),
  n_cores = 20L,
  min_land_percent = 20L,
  path_chelsa_tif = NULL,
  temp_dir = NULL,
  predictor_pattern = "bio|npp"
)
}
\arguments{
\item{agg_factors}{Integer vector. Aggregation factors for spatial
resolution. Original CHELSA data is at 30 arc-seconds (~1km at the
equator). Aggregation factors are applied to both latitude and longitude
dimensions. Defaults to \code{c(5L, 10L, 20L)} for 2.5, 5, and 10 arc-minutes}

\item{n_cores}{Integer. Number of CPU cores to use for parallel processing;
defaults to \code{20L}.}

\item{min_land_percent}{Numeric. Minimum percentage of land required in an
aggregated grid cell to be considered as land. Value must be between 0 and
100; defaults to 20L for 20\% land.}

\item{path_chelsa_tif}{Character or \code{NULL}. Path to directory containing
pre-downloaded CHELSA \code{.tif} files. If \code{NULL}, a temporary directory is
used and files are downloaded/removed as needed. If a directory path is
provided, it must exist and contain the original CHELSA \code{.tif} files.}

\item{temp_dir}{Character or \code{NULL}. Path to temporary directory for
intermediate files, including downloaded CHELSA \code{.tif} files if
\code{path_chelsa_tif} is \code{NULL} and temporary files created during processing.
If \code{NULL}, a new temporary directory is created. The directory and its
contents are deleted after processing.}

\item{predictor_pattern}{Character. Regular expression pattern that is
compatible with \link[stringr:str_detect]{stringr::str_detect} to filter CHELSA predictor variables.
See CHELSA \href{https://chelsa-climate.org/downloads/}{Technical Specifications} for a list of
available variables and the \code{var_name} column of \link[ecokit:get_chelsa_links]{ecokit::get_chelsa_links}
for the list of supported variables. Defaults to \code{"bio|npp"} for 19
bioclimatic and net primary productivity variables. Use \code{".+"} to
process all available variables.}
}
\value{
The function invisibly returns \code{NULL}. Side effects include writing
processed data to the \code{climate} sub-directory:
\itemize{
\item Summary data frame
(\code{climate/climate_data.RData}) contains metadata for each processed raster.
\item Tiff files are saved in respective sub-directories:
\verb{1981_2010_res_<agg_factor>} for historic climates (1981-2010) or future
projections with the structure:
\verb{<time_period>_<climate_scenario>_<climate_model>_res_<agg_factor>/}.
\itemize{
\item \verb{<time_period>}: \verb{2011_2040}, \verb{2041_2070}, or \verb{2071_2100}.
\item \verb{<climate_scenario>}: ssp126, ssp370, or ssp585.
\item \verb{<climate_model>}: e.g., \code{gfdl-esm4}, \code{ipsl_cm6a_lr}, \code{mpi_esm1_2_hr},
\code{mpi_esm2_0}, and \code{ukesm1_0_LL}.
\item \verb{<agg_factor>}: Aggregation factor (\code{5}, \code{10}, or \code{20}).
}
}
}
\description{
Prepare global terrestrial \href{https://chelsa-climate.org/}{CHELSA} climate data
as TIFF files at different spatial resolutions and future projections for use
in species distribution modelling.
}
\details{
\itemize{
\item \emph{This function is not intended to be called by the end user of the \code{OneSDM}
package, but to prepare necessary climate data for use in \code{OneSDM} species
distribution modelling workflow.}
\item The function loads \href{https://doi.org/10.48364/ISIMIP.836809.3/}{land masks}
and prepares reference masks at various resolutions, retrieves CHELSA data
links from \link[ecokit:get_chelsa_links]{ecokit::get_chelsa_links}, and processes climate rasters in
parallel. It handles downloading, cropping, masking, aggregating, and writing
output rasters, as well as attaching metadata.
}
}
\examples{
\dontrun{
  prepare_climate()
}

}
\author{
Ahmed El-Gabbas
}
