% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_species.R
\name{prepare_species_data}
\alias{prepare_species_data}
\title{Prepare Species Occurrence Data for Species Distribution Modelling using OneSDM}
\usage{
prepare_species_data(
  gbif_ids = NULL,
  easin_ids = NULL,
  coordinates = NULL,
  model_dir = NULL,
  climate_dir = NULL,
  resolution = NULL,
  verbose = TRUE,
  exclude_extents = list(),
  species_name = "species",
  outlier_dist_km = 0L,
  outlier_resolution = 0.125,
  outlier_n_cores = 6L,
  plot_distribution = TRUE
)
}
\arguments{
\item{gbif_ids}{\emph{(character or numeric)}. A vector of one or more GBIF
taxon keys. When multiple IDs are supplied, data are combined across all keys.
If \code{NULL} (default), the function attempts to retrieve IDs from the
"\code{onesdm_gbif_ids}" option. If no IDs are found, the GBIF download step is
skipped. To request data from GBIF, GBIF credentials must be set for the
current session. See
\href{https://docs.ropensci.org/rgbif/articles/gbif_credentials.html}{article}
from \code{rgbif} R package documentation on how to set the GBIF credentials. See
\link{prepare_gbif_data} for details. Optional.}

\item{easin_ids}{\emph{(character)} A vector of one or more EASIN species IDs,
each starting with "R" followed by five digits (e.g., "R00544"). Species IDs can
be obtained from the
\href{https://easin.jrc.ec.europa.eu/spexplorer/search/}{EASIN website}
by searching for a species and checking its EASIN ID in the species details
section. When multiple IDs are provided, data are collated across all IDs.
If \code{NULL}(default), the function attempts to retrieve IDs from the
"\code{onesdm_easin_ids}" option option and skips the EASIN download if no IDs
are found. See \link{prepare_easin_data} for details. Optional.}

\item{coordinates}{\emph{(data frame or matrix)}. A user-supplied object
containing longitude and latitude values in two columns (in that order).
Must have exactly two columns. If \code{NULL} (default), the function attempts
to retrieve coordinates from the "\code{onesdm_coordinates}" option. See
\code{\link[=prepare_user_data]{prepare_user_data()}} for details. Optional.}

\item{model_dir}{\emph{(character)}. Path to the directory where model outputs
will be saved. A subdirectory named \code{data} is automatically created within this
directory to store processed species data. When modelling multiple species,
it is recommended to use a separate directory for each run to avoid overwriting
or mixing data files. This path can also be set via the "\code{onesdm_model_dir}"
option.  Default is \code{NULL}. \strong{Required}.}

\item{climate_dir}{Character. Directory path for climate data files. Can be
set via the "\code{onesdm_climate_dir}" option. This directory is used to load
mask layers matching the specified resolution. The same directory should be
used in case of modelling multiple species to ensure consistency. Default
is \code{NULL}.}

\item{resolution}{\emph{(numeric)}. Spatial resolution used to prepare data
for analysis. Acceptable values are \strong{5, 10, or 20}, corresponding to
approximate spatial resolutions of 5 km, 10 km, and 20 km (2.5, 5, and 10
arc-minutes), respectively. This value can also be set via "\code{onesdm_resolution}"
option. Default is \code{NULL}. \strong{Required}.}

\item{verbose}{\emph{(logical)}. If \code{TRUE} (default), prints progress messages during
execution. Can also be set via the "\code{onesdm_verbose}" option. Default is \code{"TRUE"}.}

\item{exclude_extents}{\emph{(list)}. A list of \code{SpatExtent} objects (created with
\link[terra:ext]{terra::ext} function) defining geographic areas to exclude from the data.
An empty list (default) indicates that no areas are excluded. This argument can
also be set via the "\code{onesdm_exclude_extents}" option. Optional.}

\item{species_name}{\emph{(character)}. Name of the species used for labelling
outputs, such as maps. This argument does \strong{not} affect data retrieval from
\code{GBIF} or \code{EASIN}. Can be set via the "\code{onesdm_species_name}" option.
Default is \code{"species"}.}

\item{outlier_dist_km}{\emph{(numeric)}. Distance threshold (in kilometres) used
to identify spatial outliers. If set to \code{0L} (default), no outlier detection or
filtering is applied. This value can also be set via the "\code{onesdm_outlier_dist_km}"
option. See \link[ecokit:nearest_dist_sf]{ecokit::nearest_dist_sf} for details. Optional.}

\item{outlier_resolution}{\emph{(numeric)}. Spatial resolution (in degrees) used
for outlier detection calculations. This parameter is only applied when
\code{outlier_dist_km} is greater than 0. A coarser resolution can speed up computation
at the cost of spatial precision. Can also be set via the "\code{onesdm_outlier_resolution}"
option. See \link[ecokit:nearest_dist_sf]{ecokit::nearest_dist_sf} for details. Default is "\code{0.125}".}

\item{outlier_n_cores}{\emph{(integer)}. Number of CPU cores to use for parallel
outlier detection. This parameter is only applied when \code{outlier_dist_km} is
greater than 0. Can also be set via the "\code{onesdm_outlier_n_cores}" option.
Default is \code{6L}.}

\item{plot_distribution}{\emph{(logical)}. If \code{TRUE} (default), generates and
saves a JPEG map of the final species distribution data. Can also be set via
the "\code{onesdm_plot_distribution}" option.}
}
\value{
Invisibly returns path to the final processed species data raster
file ("\verb{species_data_r_<resolution>_km_PA.tif}") saved in the "\code{data}"
subdirectory of "\code{model_dir}".
}
\description{
This function prepares species occurrence data from one or more
sources — GBIF, EASIN, and user-provided data sets — for species distribution modelling.
It handles downloading, validating, integrating, and rasterising the data.
Optional steps include excluding specified geographic areas and filtering spatial
outliers. Internally, the function relies on \code{\link[=prepare_gbif_data]{prepare_gbif_data()}},
\code{\link[=prepare_easin_data]{prepare_easin_data()}}, and \code{\link[=prepare_user_data]{prepare_user_data()}} to retrieve and process data
from each respective source.
}
\details{
This function creates  a \code{data} subdirectory within \code{model_dir} saves multiple
outputs, including raw data, processed data, rasterised data, and optional
visualisations.
The function performs the following steps:
\itemize{
\item \strong{Validates input parameters}.
\item \strong{Retrieves species data} from from GBIF, EASIN, and/or user-provided coordinates
using \link{prepare_gbif_data}, \link{prepare_easin_data}, and \link{prepare_user_data}.
\itemize{
\item At least one of \code{easin_ids}, \code{gbif_ids}, or \code{coordinates} must be
provided.
\item Not all arguments of these functions are exposed via this wrapper; use the
respective options and consult the function documentation for details.
}
\item \strong{Merges data} from all sources.
\item \strong{Optionally removes spatial outliers} based on nearest neighbour distances.
\item \strong{Rasterises occurrence data} to match the specified resolution.
\item \strong{Excludes geographic extents} if specified.
\item \strong{Saves processed data and generates visualisations} if requested.
\itemize{
\item The map the rasterised species distribution overlaid on a world map.
\item Excluded areas (from \code{exclude_extents}) are highlighted on the map in red.
\item If a specific extent is used for GBIF data retrieval (via
\code{onesdm_gbif_boundaries} option in \link{prepare_gbif_data}), it is
highlighted in green.
Default argument values can also be set globally using the the \link[base:options]{base::options}
function, which allows users to avoid repeatedly specifying them in function calls.
Corresponding options for the function arguments are:
}
\item \code{onesdm_easin_ids}: for \code{easin_ids}
\item \code{onesdm_gbif_ids}: for \code{gbif_ids}
\item \code{onesdm_coordinates}: for \code{coordinates}
\item \code{onesdm_model_dir}: for \code{model_dir}
\item \code{onesdm_climate_dir}: for \code{climate_dir}
\item \code{onesdm_resolution}: for \code{resolution}
\item \code{onesdm_verbose}: for \code{verbose}
\item \code{onesdm_exclude_extents}: for \code{exclude_extents}
\item \code{onesdm_species_name}: for \code{species_name}
\item \code{onesdm_outlier_dist_km}: for \code{outlier_dist_km}
\item \code{onesdm_outlier_resolution}: for \code{outlier_resolution}
\item \code{onesdm_outlier_n_cores}: for \code{outlier_n_cores}
\item \code{onesdm_plot_distribution}: for \code{plot_distribution}
These options work in addition to those used by the internal functions called
within this wrapper.
}
}
\examples{
\dontrun{
options(
  onesdm_gbif_verbose = TRUE,
  onesdm_start_year = 1990L,
  onesdm_gbif_overwrite = FALSE,
  onesdm_gbif_boundaries = c(-170L, 170L, -80L, 80L),
  onesdm_gbif_max_uncertainty = 5L)

prepare_species_data(
  easin_ids = "R00042", gbif_ids = 2979128, coordinates = NULL,
  model_dir = "test/Acacia_karroo", climate_dir = "test/climate_data",
  resolution = 10L, verbose = TRUE,
  exclude_extents = list(terra::ext(110, 180, -50, -10)),
  species_name = "Acacia_karroo",
  outlier_dist_km = 200L, outlier_resolution = 0.25, outlier_n_cores = 4L)
}

}
\author{
Ahmed El-Gabbas
}
