% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesdm_data.R
\docType{data}
\name{sdm_model_settings}
\alias{sdm_model_settings}
\title{A list of default settings for species distribution modelling algorithms used
in \code{OneSDM}.}
\format{
A named list where each element corresponds to a specific modelling
method and contains its default control parameters.
}
\usage{
sdm_model_settings
}
\description{
A list of default settings for species distribution modelling algorithms used
in \code{OneSDM}.
}
\examples{
# View the SDM model settings
data("sdm_model_settings")
print(sdm_model_settings)
}
\keyword{datasets}
