% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesdm_data.R
\docType{data}
\name{landuse_data}
\alias{landuse_data}
\title{landuse_data Data}
\format{
\subsection{\code{landuse_data} A tibble 11 columns:}{
\itemize{
\item \strong{\code{pft_type}}: Character. The type of plant functional type (PFT); one of
"\code{original}" for the PFTs classes in the original land use data and
"\code{cross-walk}" for the grouped (cross-walked) PFTs.
\item \strong{\code{resolution}}: Integer. The resolution of the land use data: \code{5}, \code{10},
and \code{20}, for resolutions of approximately 10, 20, and 40 km (2.5, 5, and
10 arc-minutes) respectively.
\item \strong{\code{climate_scenario}}: Character. The climate scenario: "\code{current}",
"\code{ssp126}", "\code{ssp370}", and "\code{ssp585}".
\item \strong{\code{year}}: Character. The year ranges: "\verb{1981_2010}", "\verb{2011_2040}",
"\verb{2041_2070}", and "\verb{2071_2100}".
\item \strong{\code{climate_name}}: Character. The climate name: "\code{current}" and
\verb{"<climate_scenario>_<year}>"` for future options.
\item \strong{\code{pft_id}}: Integer. The PFT ID: 1-20 for the original PFTs and 1-12 for
the cross-walked PFTs.
\item \strong{\code{pft_name}}: Character. The plant functional type name.
\item \strong{\code{osf_path}}: Character. The OSF relative path to the land use data file;
e.g., "\code{current_res_5/landuse_pft_1_water.tif}".
\item \strong{\code{out_dir}}: Character. The output directory for the land use data file:
"\verb{res_<resolution>/1981_2010}" for current options and
"\verb{res_<resolution>/<year>_<climate_scenario>/}" for future options. This
would be relative to the \code{landuse_dir} argument.
\item \strong{\code{out_file}}: Character. The output file path for the land use data
files: "\verb{<out_dir>/lu_(pft|pft_cw)_<pft_id>.tif}". This would be relative
to the \code{landuse_dir} argument.
\item \strong{\code{out_file_name}}: Character. The output file name without the directory
and file extension; e.g., "\code{lu_pft_1}" and "\code{lu_pft_cw_1}".
}
}
}
\usage{
landuse_data
}
\description{
A dataset containing information about land use data files available for
download from OSF. For more details, see \code{\link[=prepare_landuse]{prepare_landuse()}}.
}
\examples{
# View the land use data
data("landuse_data")
head(landuse_data)
}
\keyword{datasets}
