% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSDM-package.R
\docType{package}
\name{OneSDM-package}
\alias{OneSDM}
\alias{OneSDM-package}
\title{OneSDM: Automated workflows for species distribution modelling under environmental change scenarios}
\description{
OneSTOP-Task-5.2 [more info!]
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marinaGolivets/OneSDM}
  \item Report bugs at \url{https://github.com/marinaGolivets/OneSDM/issues}
}

}
\author{
\strong{Maintainer}: Ahmed El-Gabbas \email{elgabbas@outlook.com} (\href{https://orcid.org/0000-0003-2225-088X}{ORCID})

Authors:
\itemize{
  \item Marina Golivetes \email{marina.golivetes@ufz.de} (\href{https://orcid.org/0000-0003-1278-2663}{ORCID})
}

}
\keyword{internal}
