% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_easin.R
\name{prepare_easin_data}
\alias{prepare_easin_data}
\title{Download and Clean EASIN Occurrence Data}
\usage{
prepare_easin_data(
  easin_ids = NULL,
  model_dir = NULL,
  timeout = 600L,
  n_search = 1000L,
  n_attempts = 10L,
  sleep_time = 5L,
  exclude_gbif = TRUE,
  verbose = TRUE,
  start_year = 1981L,
  overwrite = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{easin_ids}{\emph{(character)} A vector of one or more EASIN IDs,
each starting with "R" followed by five digits (e.g., "R00544"). EASIN IDs can
be obtained from the
\href{https://easin.jrc.ec.europa.eu/spexplorer/search/}{EASIN website}
by searching for a taxon name. When multiple EASIN IDs are provided, data
are downloaded for each ID and then collated.
If \code{NULL} (default), the function attempts to retrieve IDs from the
"\code{onesdm_easin_ids}" option and skips the EASIN download if no IDs
are found. \strong{Required}.}

\item{model_dir}{\emph{(character)} Path to the directory where model outputs
will be saved. A subdirectory named \code{data} is automatically created within this
directory to store processed taxon data. When modelling multiple taxa,
it is recommended to use a separate directory for each run to avoid overwriting
or mixing data files. This path can also be set via the \code{onesdm_model_dir}
option.  Default is \code{NULL}.}

\item{timeout}{\emph{(integer)} Timeout (in seconds) for each download attempt.
Default is \code{600L}. Can also be set via the "\code{onesdm_easin_timeout}" option.
This timeout applies to each download chunk separately and does not
limit the total duration of the full download process.}

\item{n_search}{\emph{(integer)} Number of records requested per API call (chunk
size). Default is \code{1000L}, which is the maximum allowed by the EASIN API.
Can also be set via the "\code{onesdm_easin_n_search}" option.}

\item{n_attempts}{\emph{(integer)} Maximum number of download attempts per chunk.
Default is \code{10L}. Can also be set via the "\code{onesdm_easin_n_attempts}"
option.}

\item{sleep_time}{\emph{(integer)} Time to wait (in seconds) between successive
chunk downloads. This delay helps prevent overloading the EASIN server.
Default is \code{5L}. Can also be set via the "\code{onesdm_easin_sleep_time}" option.}

\item{exclude_gbif}{\emph{(logical)} If \code{TRUE} (default),
\href{https://www.gbif.org/}{GBIF} records are omitted from the download.
Can also be set via the "\code{onesdm_easin_exclude_gbif}" option.}

\item{verbose}{\emph{(logical)} If \code{TRUE} (default), progress and information
messages are printed, including the URL of the currently processed chunk.
Can also be set via the "\code{onesdm_easin_verbose}" option.}

\item{start_year}{\emph{(integer)} The starting year from which records are included.
The default is \code{1981L}, to match the temporal coverage of CHELSA climate data.
Can also be set via the "\code{onesdm_start_year}" option.}

\item{overwrite}{\emph{(logical)} If \code{TRUE}, overwrites existing cleaned EASIN data
in the modelling directory. Default is \code{FALSE}. Can also be set via
"\code{onesdm_easin_overwrite}" option.}

\item{return_data}{\emph{(logical)} If \code{TRUE}, returns the processed EASIN data as an
\code{sf} object in addition to saving it. Default is \code{FALSE}.}
}
\value{
Invisibly returns the file path to the saved cleaned EASIN data
\code{easin_data.RData} file in the \code{data} subdirectory of \code{model_dir}, unless
\code{return_data} is \code{TRUE}, in which case it returns the cleaned EASIN data as
an \code{sf} object.
}
\description{
Downloads, combines, cleans, and saves taxon occurrence data for given
EASIN IDs from the \href{https://easin.jrc.ec.europa.eu/easin}{EASIN}
(European Alien Species Information Network) Geodatabase.
}
\details{
\itemize{
\item The function supports chunked downloads, multiple attempts, and
extensive data cleaning, including coordinate precision checks and spatial
filtering using the \code{CoordinateCleaner} package. The cleaned data is saved as
an \code{.RData} file in the specified model directory.
\item The function checks for an existing cleaned EASIN data file and skips
download if found.
\item The function applies several cleaning steps; e.g. filtering out records
with low coordinate precision, equal longitude / latitude, or near centroids,
capitals, biodiversity institutions, and GBIF headquarters.
\item Function default arguments can be set globally using the \link[base:options]{base::options}
function. Users can set these options at the start of their R session to
avoid repeatedly specifying them in function calls. The following options
correspond to the function arguments:
\itemize{
\item "\code{onesdm_easin_ids}": Character vector of EASIN taxon IDs.
\item "\code{onesdm_model_dir}": Character. Path to the modelling directory.
\item "\code{onesdm_easin_timeout}": Integer. Timeout (in seconds) for each download
attempt.
\item "\code{onesdm_easin_n_search}": Integer. Number of records to request per API
call.
\item "\code{onesdm_easin_n_attempts}": Integer. Maximum number of download attempts
per chunk.
\item "\code{onesdm_easin_sleep_time}": Integer. Seconds to wait between chunk
downloads.
\item "\code{onesdm_easin_exclude_gbif}": Logical. Whether to exclude GBIF records.
\item "\code{onesdm_easin_verbose}": Logical. Whether to print progress messages.
\item "\code{onesdm_start_year}": Integer. Minimum year for records to include.
\item "\code{onesdm_easin_overwrite}": Logical. Whether to overwrite existing
cleaned data.
}
\item Example of setting options:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(
  onesdm_easin_ids = c("R00042", "R00544"),
  onesdm_easin_exclude_gbif = TRUE,
  onesdm_easin_overwrite = TRUE,
  onesdm_model_dir = "path/to/model_dir"
)
}\if{html}{\out{</div>}}
}
}
\examples{
\dontrun{
 require(fs)

 # Prepare EASIN data for species with EASIN ID "R00042": Acacia karroo Hayne
 temp_model_dir <- fs::path_temp("onesdm_model_dir")
 easin_data <- prepare_easin_data(
   easin_ids = "R00042", model_dir = temp_model_dir, return_data = TRUE)

 print(easin_data)

 # # ||||||||||||||||||||||||||||||||||||||||||||||||||| #

 # Prepare EASIN data, using argument values from `options`
 temp_model_dir_2 <- fs::path_temp("onesdm_model_dir")
 options(
    onesdm_easin_ids = c("R00042", "R00544"),
    onesdm_easin_exclude_gbif = TRUE,
    onesdm_easin_overwrite = TRUE,
    onesdm_model_dir = temp_model_dir_2)

 prepare_easin_data()

}

}
\references{
EASIN Geospatial Web Service:
\url{https://easin.jrc.ec.europa.eu/apixg/home/geoqueries/}
}
\author{
Ahmed El-Gabbas, Maryna Golivets
}
