% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_easin.R
\name{prepare_easin_data}
\alias{prepare_easin_data}
\title{Download and Clean EASIN Occurrence Data for Given Species IDs}
\usage{
prepare_easin_data(
  easin_ids = NULL,
  model_dir = NULL,
  timeout = 300L,
  n_search = 1000L,
  n_attempts = 10L,
  sleep_time = 5L,
  exclude_gbif = TRUE,
  verbose = TRUE,
  start_year = 1981L,
  overwrite = FALSE
)
}
\arguments{
\item{easin_ids}{Character vector. One or more EASIN species IDs. Each ID
should start with 'R' followed by five digits (e.g. "R00544"). This cannot
be \code{NULL}. Species IDs can be found on the \href{https://easin.jrc.ec.europa.eu/spexplorer/search/}{EASIN website} by searching
for a species and locating the "EASIN ID" in the species details section.
Can also be set via the \code{onestop_easin_ids} option.}

\item{model_dir}{Character. Path to the modelling directory where data and
fitted models will be saved. This cannot be \code{NULL} and must the same
directory used for the same species data. This can also be set via the
\code{onestop_model_dir} option.}

\item{timeout}{Integer. Timeout (in seconds) for each download attempt.
Default is 300L. Can also be set via the \code{onestop_easin_timeout} option.
Note: This timeout only applies to the download of each chunk, not the
entire download process.}

\item{n_search}{Integer. Number of records to request per API call (chunk
size). Default is 1000L, which is the maximum allowed by the EASIN API. Can
also be set via the \code{onestop_easin_n_search} option.}

\item{n_attempts}{Integer. Maximum number of download attempts per chunk.
Default is 10L. Can also be set via the \code{onestop_easin_n_attempts} option.}

\item{sleep_time}{Integer. Seconds to wait between chunk downloads. This
helps avoid overwhelming the EASIN server. Default is 5L. Can also be set
via the \code{onestop_easin_sleep_time} option.}

\item{exclude_gbif}{Logical. If \code{TRUE} (default), exclude
\href{https://www.gbif.org/}{GBIF} — the Global Biodiversity Information
Facility — records from the download. Can also be set via the
\code{onestop_easin_exclude_gbif} option.}

\item{verbose}{Logical. If \code{TRUE} (default), print progress and information
messages, including the URL of the currently processed chunk. Can also be
set via the \code{onestop_easin_verbose} option.}

\item{start_year}{Integer. Include only records from this year onward. The
default is 1981L. Can also be set via the \code{onestop_start_year} option.}

\item{overwrite}{Logical. If \code{TRUE}, overwrite existing cleaned EASIN data
file in the model directory. Default is \code{FALSE}. Can also be set via
\code{onestop_easin_overwrite} option.}
}
\value{
(Invisibly) the path to the saved \code{.RData} file containing the
cleaned EASIN data (\code{data/easin_data.RData}) in the specified \code{model_dir}.
The function saves an empty tibble if no cleaned data is available after
processing.
}
\description{
Downloads, combines, cleans, and saves occurrence data for given
EASIN species IDs from the \href{https://alien.jrc.ec.europa.eu/easin}{EASIN}
(European Alien Species Information Network) API.
}
\details{
\itemize{
\item The function supports chunked downloads, multiple attempts, and
extensive data cleaning, including coordinate precision checks and spatial
filtering using the \code{CoordinateCleaner} package. The cleaned data is saved as
an \code{.RData} file in the specified model directory.
\item The function checks for an existing cleaned EASIN data file and skips
download if found.
\item The function applies several cleaning steps; e.g. filtering out records
with low coordinate precision, equal longitude / latitude, or near centroids,
capitals, biodiversity institutions, and GBIF headquarters.
}
}
\examples{
\dontrun{
  prepare_easin_data(easin_ids = "R00042", model_dir = "path/to/model_dir")
}

}
\references{
EASIN geospatial Web service:
\url{https://easin.jrc.ec.europa.eu/apixg/home/geoqueries/}
}
\author{
Ahmed El-Gabbas
}
